move saveBackgroundImage to runtime
