+++ src/main/deltachat/controller.ts
@@ -8,7 +8,6 @@ import DCChat from './chat'
 import DCContext from './context'
 import DCLoginController from './login'
 import DCMessageList from './messagelist'
-import DCSettings from './settings'
 import DCStickers from './stickers'
 import { ExtendedAppMainProcess } from '../types'
 import { Context } from 'deltachat-node/node/dist/context'
@@ -239,7 +238,6 @@ export default class DeltaChatController extends EventEmitter {
   readonly chat = new DCChat(this)
   readonly login = new DCLoginController(this)
   readonly messageList = new DCMessageList(this)
-  readonly settings = new DCSettings(this)
   readonly stickers = new DCStickers(this)
   readonly context = new DCContext(this)
   readonly webxdc = new DCWebxdc(this)
