+++ src/renderer/runtime.ts
@@ -122,6 +122,7 @@ interface Runtime {
     data: string
   } | null>
   resolveThemeAddress(address: string): Promise<string>
+  saveBackgroundImage(file: string, isDefaultPicture: boolean): Promise<string>
 }
 
 class Browser implements Runtime {
@@ -129,6 +130,12 @@ class Browser implements Runtime {
     throw new Error('Method not implemented.')
   }
   notifyWebxdcInstanceDeleted(_accountId: number, _instanceId: number): void {
+    throw new Error('Method not implemented.')
+  }
+  saveBackgroundImage(
+    _file: string,
+    _isDefaultPicture: boolean
+  ): Promise<string> {
     throw new Error('Method not implemented.')
   }
   getLocaleData(_locale?: string | undefined): Promise<LocaleData> {
@@ -263,6 +270,12 @@ class Electron implements Runtime {
   }
   notifyWebxdcInstanceDeleted(accountId: number, instanceId: number): void {
     ipcBackend.invoke('webxdc:instance-deleted', accountId, instanceId)
+  }
+  saveBackgroundImage(
+    file: string,
+    isDefaultPicture: boolean
+  ): Promise<string> {
+    return ipcBackend.invoke('saveBackgroundImage', file, isDefaultPicture)
   }
   getLocaleData(locale?: string | undefined): Promise<LocaleData> {
     return ipcBackend.invoke('getLocaleData', locale)
